webhooks/x-hashtag-notifier/requirements.txt
requests==2.25.1
python-dotenv==1.1.0